重构添加浏览记录的代码逻辑。

+++ cloud-collect-core/src/main/kotlin/com/windea/demo/cloudcollect/core/service/impl/HistoryServiceImpl.kt
@@ -16,7 +16,11 @@ class HistoryServiceImpl(
 	@Transactional
 	@CacheEvict(allEntries = true)
 	override fun create(history: History) {
-		//首先要删除之前的相同收藏的浏览记录
+		//相对于最后浏览记录，不要重复添加浏览记录
+		val lastHistory = historyRepository.findFirstByUserIdOrderByIdDesc(history.user.id)
+		if(history.collect.id == lastHistory.collect.id) return
+		
+		//首先要删除之前的重复的浏览记录
 		historyRepository.deleteByCollectIdAndUserId(history.collect.id, history.user.id)
 		
 		historyRepository.save(history)
